[physics] Fix RayCollection.contains raycast (again)
The problem was that in some situations we could choose a initial point
(pIni) that wouldn't cross a proper edge of the polygon. It could cross
a null adjacency segment (due to an infinite ray) making the algorithm
understand that the point is outside the polygon.

We fix that by always choosing the RayCollection.lightPos as initial
point of the raycast. It is usually outside the polygon (we need to fix
that) and crosses valid segments (the adjacency segment between the
first and last rays are always valid).

Signed-off-by: Mailson Menezes <mailson.menezes@openbossa.org>