[physics] Split algorithm to compute rays in 2 steps
The first step is to generate Rays that comes out of the object ends. No
matter if the rays would collide with another object.

After that we take all generated rays and compute whether it should be
decreased (due to collision) or removed (due to occlusion).

This fixes the issue we had when objects far from the light source was
evaluated earlier than objects closer to the light source.

Signed-off-by: Mailson Menezes <mailson.menezes@openbossa.org>