[physics] Fix occlusion bug
When a opaque object is between the lightPos and the starting point of a
ray, it means that that Ray is occluded by the opaque object and should
be removed. However if the opaque object is behind the starting line
segment of the RayCollection (the segment between the first and the last
point of the collection), no occlusion should happen.

Signed-off-by: Mailson Menezes <mailson.menezes@openbossa.org>